zaius_vcs.sh: Only rebind ucd9000 if we unbound it
Merge from Witherspoon

This makes way for an incoming kernel patch affecting
Witherspoon[1], where we disable the UCD90160 device
in the devicetree. If the driver hasn't been bound to
the device we can immediately proceed to invoking
i2c{g,s}et. Finally if it wasn't bound before the script
was executed then don't try to rebind it, as it was
disabled for a reason.

Change-Id: I6e1b6e4518f1bbfc1b29bd0fa9566d28025b9d84
Signed-off-by: John Wang <hsienchiang@gmail.com>