meta-zaius: Change TimeOwner to SPLIT
The comment of SPLIT in
https://github.com/openbmc/phosphor-time-manager/blob/master/time-config.hpp#L31.

Both BMC and HOST will maintain their individual clocks but
then the time information is stored in BMC. BMC can have
either NTP or MANUAL as it's source of time and will set
the time directly on the BMC. When IPMI_SET_SEL_TIME is
received, then the delta between that and BMC's time is
calculated and is stored.
When BMC reads the time, the current time is returned.
When IPMI_GET_SEL_TIME is received, BMC's time is retrieved
and then the delta offset is factored in prior to returning.

Change-Id: I2285d782d04786c4408ac03d74cc0a886cc05053
Signed-off-by: John Wang <hsienchiang@gmail.com>